构建：合并vcpkg静态依赖统一

+++ kbe/src/lib/entitydef/scriptdef_module.cpp
@@ -487,7 +487,7 @@ bool ScriptDefModule::addPropertyDescription(const char* attrName,
 	if(hasMethodName(attrName))
 	{
 		ERROR_MSG(fmt::format("ScriptDefModule::addPropertyDescription: There is a method[{}] name conflict! componentType={}.\n",
-			attrName, componentType));
+			attrName, static_cast<int>(componentType)));
 		
 		return false;
 	}
@@ -527,7 +527,7 @@ bool ScriptDefModule::addPropertyDescription(const char* attrName,
 	if(f_propertyDescription)
 	{
 		ERROR_MSG(fmt::format("ScriptDefModule::addPropertyDescription: [{}] is exist! componentType={}.\n",
-			attrName, componentType));
+			attrName, static_cast<int>(componentType)));
 
 		return false;
 	}
